pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "root-mcp"
version = "0.1.2"
description = "Production-grade MCP server for CERN ROOT file analysis"
authors = [{name = "Mohamed Elashri"}]
maintainers = [{name = "Mohamed Elashri"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["ROOT", "CERN", "HEP", "MCP", "particle physics"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Physics",
]

dependencies = [
    "uproot>=5.0.0,<6.0.0",
    "awkward>=2.0.0,<3.0.0",
    "numpy>=1.24.0,<2.0.0",
    "pandas>=2.0.0,<3.0.0",
    "mcp>=0.9.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=23.0.0",
    "fsspec>=2023.0.0",
    "pyarrow>=14.0.0",
    "pyyaml>=6.0",
    "xxhash>=3.0.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "ty>=0.0.2",
    "types-PyYAML>=6.0.0.0",
    "pre-commit>=3.0.0",
    "build>=1.0.0",
    "twine>=4.0.0",
]

xrootd = [
    "XRootD>=5.5.0",
]

[project.scripts]
root-mcp = "root_mcp.server:main"

[project.urls]
Homepage = "https://github.com/MohamedElashri/root-mcp"
Repository = "https://github.com/MohamedElashri/root-mcp"
Issues = "https://github.com/MohamedElashri/root-mcp/issues"
Documentation = "https://github.com/MohamedElashri/root-mcp/tree/main/docs"
Changelog = "https://github.com/MohamedElashri/root-mcp/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ty.environment]
root = ["./src"]

[tool.ty.rules]
unused-ignore-comment = "warn"
redundant-cast = "ignore"
possibly-missing-attribute = "error"
possibly-missing-import = "error"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
